requirements.txt
GitPython~=3.1.27
python-gitlab~=3.4.0
